Removed checks for updates
