+++ .pre-commit-config.yaml
@@ -77,7 +77,7 @@ repos:
         language: system
         files: setup.py
   - repo: https://github.com/astral-sh/ruff-pre-commit
-    rev: v0.12.9
+    rev: v0.12.10
     hooks:
       - id: ruff
         exclude: '__pycache__/'
