+++ albumentations/__init__.py
@@ -10,10 +10,7 @@ except Exception:  # noqa: BLE001
     __author__ = "Vladimir Iglovikov"
     __maintainer__ = "Vladimir Iglovikov"
 
-import os
 from contextlib import suppress
-
-from albumentations.check_version import check_for_updates
 
 from .augmentations import *
 from .core.composition import *
@@ -22,7 +19,3 @@ from .core.transforms_interface import *
 
 with suppress(ImportError):
     from .pytorch import *
-
-# Perform the version check after all other initializations
-if os.getenv("NO_ALBUMENTATIONS_UPDATE", "").lower() not in {"true", "1"}:
-    check_for_updates()
